+++ Greedy Algorithms/Bellman Ford/Bellman-Ford.py
