+++ Searching and Sorting/Binary Search/Python/BinarySearch.py
