+++ Searching and Sorting/Exponential Search/C++/Exponential Search.cpp
