+++ Searching and Sorting/Quick Sort/Python/quick sort.py
